reorganise spec structure
