+++ lib/getDependencyModules.js
@@ -2,20 +2,20 @@
 
 const debug = require('debug');
 const stat = require('fs').statSync;
+const resolvePath = require('path').resolve;
 const includes = require('lodash/includes');
 const d = debug('breadboard:setup');
 
 module.exports = (packageDir, blacklist, substitutes) => {
-  const packageJsonPath = `${packageDir}/package.json`;
+  const absolutePackageJsonPath = resolvePath(process.cwd(), `${packageDir}/package.json`);
 
   return new Promise((resolve, reject) => {
     let packageJsonModules = {};
     let packageJsonModuleNames;
 
     try {
-      stat(packageJsonPath);
-      packageJsonModuleNames = Object.keys(require(packageJsonPath).dependencies);
-
+      stat(absolutePackageJsonPath);
+      packageJsonModuleNames = Object.keys(require(absolutePackageJsonPath).dependencies);
       d('package.json modules', Object.keys(packageJsonModules));
     }
     catch (e) {
