+++ spec/lib/getNativeModule.spec.js
@@ -1,11 +1,11 @@
 const test = require('ava');
 const sinon = require('sinon');
-const subject = require('../lib/getNativeModules');
+const subject = require('../../lib/getNativeModules');
 const sandbox = sinon.sandbox.create();
 const internalModule = 'internal/fakeNativeModule';
 const nonInternalModule = '../spec/fixtures/fakeNativeModule';
 const modules = {
-  '../spec/fixtures/fakeNativeModule': require('../spec/fixtures/fakeNativeModule'),
+  '../spec/fixtures/fakeNativeModule': require('../fixtures/fakeNativeModule'),
   'internal/fakeNativeModule': {}
 };
 const bindingStub = sandbox.stub(process, 'binding');
@@ -18,7 +18,7 @@ test.after(() => {
   sandbox.restore();
 });
 test('non-internal native modules are returned', async t => {
-  const fakeNonInternalModule = require('./fixtures/fakeNativeModule');
+  const fakeNonInternalModule = require('../fixtures/fakeNativeModule');
   const nativeModules = await subject([]);
 
   t.same(nativeModules[nonInternalModule], fakeNonInternalModule);
