+++ spec/lib/loadModule.spec.js
@@ -1,12 +1,12 @@
 const test = require('ava');
-const subject = require('../lib/loadModule');
+const subject = require('../../lib/loadModule');
 const moduleDirectory = '../spec/fixtures';
 
 test('creates a getter for a CommonJS module', t => {
   const fileStat = {
     name: 'fakeModule.js'
   };
-  const expectedModule = require('./fixtures/fakeModule');
+  const expectedModule = require('../fixtures/fakeModule');
   const module = subject(moduleDirectory, fileStat)();
   t.same(module, expectedModule);
 });
@@ -14,7 +14,7 @@ test('creates a getter for a JSON file', t => {
   const fileStat = {
     name: 'fakeConfig.json'
   };
-  const expectedModule = require('./fixtures/fakeConfig.json');
+  const expectedModule = require('../fixtures/fakeConfig.json');
   const module = subject(moduleDirectory, fileStat)();
   t.same(module, expectedModule);
 });
